toolkit/components/extensions/test/mochitest/test_ext_localStorage.html

<!DOCTYPE HTML>
<html>
<head>
  <title>WebExtension test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

function backgroundScript() {
  let hasRun = localStorage.getItem("has-run");
  let result;
  if (!hasRun) {
    localStorage.setItem("has-run", "yup");
    localStorage.setItem("test-item", "item1");
    result = "item1";
  } else {
    let data = localStorage.getItem("test-item");
    if (data == "item1") {
      localStorage.setItem("test-item", "item2");
      result = "item2";
    } else if (data == "item2") {
      localStorage.removeItem("test-item");
      result = "deleted";
    } else if (!data) {
      localStorage.clear();
      result = "cleared";
    }
  }
  browser.test.sendMessage("result", result);
  browser.test.notifyPass("localStorage");
}

let extensionData = {
  background: "(" + backgroundScript.toString() + ")()",
};

add_task(function* test_contentscript() {
  let id = "test-webextension@mozilla.com";
  const RESULTS = ["item1", "item2", "deleted", "cleared", "item1"];

  for (let expected of RESULTS) {
    let extension = ExtensionTestUtils.loadExtension(extensionData, id);
    let [, actual] = yield Promise.all([extension.startup(), extension.awaitMessage("result")]);
    yield extension.awaitFinish("localStorage");
    yield extension.unload();

    is(actual, expected, "got expected localStorage data");
  }
});
</script>

</body>
</html>
